requirements.txt
pandas
numpy
google-generativeai
black
flake8
isort
loguru
pip
python-dotenv
tqdm
typer
-e .
